pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "printer-maintenance-manager"
version = "1.0.0"
description = "Moonraker plugin for printer usage tracking and maintenance management"
readme = "README.md"
license = { text = "MIT" }
authors = [
  { name = "Your Name", email = "you@email.com" }
]
requires-python = ">=3.8"

dependencies = [
  "PyYAML",
  "requests"
]

[project.urls]
Homepage = "https://github.com/harrynow83/printer-maintenance-manager"
Repository = "https://github.com/harrynow83/printer-maintenance-manager"
Issues = "https://github.com/harrynow83/printer-maintenance-manager/issues"
